Update README.md
jira-gerrit-plugin didn't find any change (e.g. tr:%s, message:%s) until I've added the JIRA user in the "Non-Interactive Users" group on Gerrit. That behavior has been seen on Gerrit 2.11.3

At "Features" section there is a topic[1] that gives a clue regarding that, but that should be present in the "Getting started" section as well

[1] * The Gerrit user configured in the admin panel must have access to perform
    all necessary steps